+++ contracts/interface/IExchange.sol
@@ -23,8 +23,8 @@ interface IExchange {
         int256 exchangedPositionSize;
         int256 exchangedPositionNotional;
         uint256 fee;
-        int256 openNotional;
-        int256 realizedPnl;
+        uint256 insuranceFundFee;
+        int256 pnlToBeRealized;
         uint256 sqrtPriceAfterX96;
         int24 tick;
         bool isPartialClose;
@@ -45,9 +45,6 @@ interface IExchange {
         int256 deltaAvailableQuote;
     }
 
-    /// @param fundingPayment > 0: payment, < 0 : receipt
-    event FundingPaymentSettled(address indexed trader, address indexed baseToken, int256 fundingPayment);
-
     event FundingUpdated(address indexed baseToken, uint256 markTwap, uint256 indexTwap);
 
     event MaxTickCrossedWithinBlockChanged(address indexed baseToken, uint24 maxTickCrossedWithinBlock);
@@ -59,7 +56,7 @@ interface IExchange {
 
     function settleFunding(address trader, address baseToken)
         external
-        returns (Funding.Growth memory fundingGrowthGlobal);
+        returns (int256 fundingPayment, Funding.Growth memory fundingGrowthGlobal);
 
     function getMaxTickCrossedWithinBlock(address baseToken) external view returns (uint24);
 
@@ -91,6 +88,4 @@ interface IExchange {
     function getAccountBalance() external view returns (address);
 
     function getClearingHouseConfig() external view returns (address);
-
-    function getInsuranceFund() external view returns (address);
 }