+++ test/clearingHouse/fixtures.ts
@@ -112,17 +112,12 @@ export function createClearingHouseFixture(
         await insuranceFund.initialize(USDC.address)
 
         // deploy exchange
-        await exchange.initialize(
-            marketRegistry.address,
-            orderBook.address,
-            clearingHouseConfig.address,
-            insuranceFund.address,
-        )
+        await exchange.initialize(marketRegistry.address, orderBook.address, clearingHouseConfig.address)
         exchange.setAccountBalance(accountBalance.address)
 
         await orderBook.setExchange(exchange.address)
 
-        await accountBalance.initialize(clearingHouseConfig.address, exchange.address)
+        await accountBalance.initialize(clearingHouseConfig.address, orderBook.address)
 
         const vaultFactory = await ethers.getContractFactory("Vault")
         const vault = (await vaultFactory.deploy()) as Vault
@@ -164,6 +159,7 @@ export function createClearingHouseFixture(
                 uniV3Factory.address,
                 exchange.address,
                 accountBalance.address,
+                insuranceFund.address,
             )
         } else {
             const clearingHouseFactory = await ethers.getContractFactory("ClearingHouse")
@@ -175,6 +171,7 @@ export function createClearingHouseFixture(
                 uniV3Factory.address,
                 exchange.address,
                 accountBalance.address,
+                insuranceFund.address,
             )
         }
 
@@ -308,12 +305,7 @@ export async function mockedClearingHouseFixture(): Promise<MockedClearingHouseF
 
     const exchangeFactory = await ethers.getContractFactory("Exchange")
     const exchange = (await exchangeFactory.deploy()) as Exchange
-    await exchange.initialize(
-        mockedMarketRegistry.address,
-        mockedOrderBook.address,
-        clearingHouseConfig.address,
-        insuranceFund.address,
-    )
+    await exchange.initialize(mockedMarketRegistry.address, mockedOrderBook.address, clearingHouseConfig.address)
     const mockedExchange = await smockit(exchange)
 
     const accountBalanceFactory = await ethers.getContractFactory("AccountBalance")
@@ -335,6 +327,7 @@ export async function mockedClearingHouseFixture(): Promise<MockedClearingHouseF
         mockedUniV3Factory.address,
         mockedExchange.address,
         mockedAccountBalance.address,
+        insuranceFund.address,
     )
     return {
         clearingHouse,